+++ docs/containers/SideNavContainer/index.js
@@ -1,4 +1,5 @@
 import React from 'react';
+
 import components from 'docs/data/index.json';
 import SideNav from 'docs/components/SideNav/';
 import { toTitleCase } from 'docs/utils/';
@@ -12,7 +13,7 @@ const SideNavContainer = ({ categoryName, currentRoute }) => {
     if (!component.private) {
       sideNavItems.push({
         name: component.name,
-        href: `/${categoryName}/${component.name}/`,
+        href: `/${categoryName}/${component.name}/sass`,
       });
     }
   });