+++ src/components/Input/index.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import Label from '../Label';
 import classnames from 'classnames';
+import PropTypes from 'prop-types';
 
 /**
  * Generates an `input` element (optionally wrapped in a label) and accepts
@@ -8,7 +9,9 @@ import classnames from 'classnames';
  * @param {Object} props - Properties passed to component
  * @returns {ReactElement}
  */
+
 class Input extends React.Component {
+
   blur() {
     if (this._input) {
       this._input.blur();
@@ -16,61 +19,67 @@ class Input extends React.Component {
   }
 
   renderInput(opts) {
-    let classes = classnames('oui-text-input', {'oui-text-input--search': opts.isFilter}, {'oui-form-bad-news': opts.hasError});
+    let noteLabel = null;
+    if (opts.note) {
+      noteLabel = <div className="oui-form-note">{ opts.note }</div>;
+    }
+
+    let wrapperClasses = classnames(
+      {'oui-form-bad-news': opts.displayError}
+    );
+
+    let classes = classnames(
+      'oui-text-input', {'oui-text-input--search': opts.isFilter}
+    );
 
     return (
       /* eslint-disable react/jsx-no-bind */
-      <input
-        className={ classes }
-        ref={ (c) => { this._input = c; } }
-        type={ opts.type }
-        value={ opts.value }
-        defaultValue={ opts.defaultValue }
-        placeholder={ opts.placeholder }
-        required={ opts.isRequired }
-        readOnly={ opts.isReadOnly }
-        disabled={ opts.isDisabled }
-        onInput={ opts.onInput }
-        onChange={ opts.onChange }
-        onBlur={ opts.onBlur }
-        onKeyDown={ opts.onKeyDown }
-        onFocus={ opts.onFocus }
-        min={ opts.min }
-        max={ opts.max }
-        data-test-section={ opts.testSection }
-      />
+      <div className={ wrapperClasses }>
+        <Label
+          displayError={ opts.displayError }
+          isRequired={ opts.isRequired }
+          isOptional={ opts.isOptional }>
+          { opts.label }
+        </Label>
+        <input
+          className={ classes }
+          ref={ (c) => { this._input = c; } }
+          type={ opts.type }
+          value={ opts.value }
+          defaultValue={ opts.defaultValue }
+          placeholder={ opts.placeholder }
+          required={ opts.isRequired }
+          readOnly={ opts.isReadOnly }
+          disabled={ opts.isDisabled }
+          onInput={ opts.onInput }
+          onChange={ opts.onChange }
+          onBlur={ opts.onBlur }
+          onKeyDown={ opts.onKeyDown }
+          onFocus={ opts.onFocus }
+          min={ opts.min }
+          max={ opts.max }
+          data-test-section={ opts.testSection }
+        />
+        { noteLabel }
+      </div>
       /* eslint-enable */
     );
   }
 
   render() {
-    if (this.props.label) {
-      return (
-        <div className={ classnames({'oui-form-bad-news': this.props.hasError}) }>
-          <Label testSection={ this.props.testSection && this.props.testSection + '-label' }>
-            <div className="oui-label">
-              { this.props.label }
-              { this.props.isOptional && <span className="oui-label__optional">(Optional)</span> }
-            </div>
-            { this.renderInput(this.props) }
-          </Label>
-        </div>
-      );
-    }
-
     return this.renderInput(this.props);
   }
 }
 
 Input.propTypes = {
   /** The default value of the input used on initial render */
-  defaultValue: React.PropTypes.string,
+  defaultValue: PropTypes.string,
   /** Toggle error state styles  */
-  hasError: React.PropTypes.bool,
+  displayError: PropTypes.bool,
   /** Prevents input from being modified and appears disabled */
-  isDisabled: React.PropTypes.bool,
-  /** Includes search icon if true */
-  isFilter: React.PropTypes.bool,
+  isDisabled: PropTypes.bool,
+  /** Includes error if true */
+  isFilter: PropTypes.bool,
   /** Adds an optional label if there is a label provided
    *  @param {Object} props Object of props
    *  @returns {Error} Error or null
@@ -82,38 +91,40 @@ Input.propTypes = {
     return null;
   },
   /** Prevents input from being modified but doesn't appear disabled */
-  isReadOnly: React.PropTypes.bool,
+  isReadOnly: PropTypes.bool,
   /** Prevents input from being submitted without value */
-  isRequired: React.PropTypes.bool,
+  isRequired: PropTypes.bool,
   /** Text that describes the input */
-  label: React.PropTypes.string,
+  label: PropTypes.string,
   /**
    * Max value for the `input`. Should be used only when `type` is `number`.
    */
-  max: React.PropTypes.number,
+  max: PropTypes.number,
   /**
    * Min value for the `input`. Should be used only when `type` is `number`.
    */
-  min: React.PropTypes.number,
+  min: PropTypes.number,
+  /** Append note near form input. */
+  note: PropTypes.string,
   /**
    * Function that fires when the input loses focus. It fires regardless of
    * whether the value has changed.
   */
-  onBlur: React.PropTypes.func,
+  onBlur: PropTypes.func,
   /** Function that fires when the input loses focus after the value changes */
-  onChange: React.PropTypes.func,
+  onChange: PropTypes.func,
   /** Function that fires when the input gains focus */
-  onFocus: React.PropTypes.func,
+  onFocus: PropTypes.func,
   /** Function that fires on keypress */
-  onInput: React.PropTypes.func,
+  onInput: PropTypes.func,
   /** Function that fires when a key is pressed down */
-  onKeyDown: React.PropTypes.func,
+  onKeyDown: PropTypes.func,
   /** Input placeholder text */
-  placeholder: React.PropTypes.string,
+  placeholder: PropTypes.string,
   /** Hook for automated JavaScript tests */
-  testSection: React.PropTypes.string,
+  testSection: PropTypes.string,
   /** Supported input types */
-  type: React.PropTypes.oneOf([
+  type: PropTypes.oneOf([
     'text',
     'password',
     'date',
@@ -124,7 +135,7 @@ Input.propTypes = {
     'tel',
   ]).isRequired,
   /** Text within the input */
-  value: React.PropTypes.string,
+  value: PropTypes.string,
 };
 
 export default Input;